Rename RegistrationUserID to RegisterID
