+++ src/App.tsx
@@ -8,7 +8,7 @@ import EditEmail from "./pages/EditEmail";
 import EditPrevPassword from "./pages/EditPrevPassword";
 import ProfileScreen from "./pages/ProfileScreen";
 import EditNewPassword from "./pages/EditNewPassword";
-import RegistrationUserID from "./pages/RegistrationUserID";
+import RegisterID from "./pages/RegisterID";
 import WelcomeScreen from "./pages/WelcomeScreen";
 
 function App() {
@@ -24,7 +24,7 @@ function App() {
                 <Route path="/profile/edit-email" element={<EditEmail/>}/>
                 <Route path="/profile/edit-prev-password" element={<EditPrevPassword/>}/>
                 <Route path="/profile/edit-new-password" element={<EditNewPassword/>}/>
-                <Route path="/registration/user-id" element={<RegistrationUserID/>}/>
+                <Route path="/registration/user-id" element={<RegisterID/>}/>
                 <Route path="/welcome" element={<WelcomeScreen/>}/>
                 <Route path="*" element={<WelcomeScreen/>}/>
             </Routes>
